# backend/openapi_pretty.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Questionnaire Agent",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/questionnaires/": {
            "post": {
                "tags": [
                    "questionnaires"
                ],
                "summary": "Upload Questionnaire",
                "description": "Upload a questionnaire to a project.",
                "operationId": "upload_questionnaire_api_v1_questionnaires__post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_questionnaire_api_v1_questionnaires__post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questionnaire"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/questionnaires/{questionnaire_id}": {
            "delete": {
                "tags": [
                    "questionnaires"
                ],
                "summary": "Delete Questionnaire",
                "description": "Delete a Questionnaire.",
                "operationId": "delete_questionnaire_api_v1_questionnaires__questionnaire_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "questionnaire_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Questionnaire Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questionnaire"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_questionnaire_api_v1_questionnaires__post": {
                "properties": {
                    "project_id": {
                        "type": "integer",
                        "title": "Project Id"
                    },
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "project_id",
                    "file"
                ],
                "title": "Body_upload_questionnaire_api_v1_questionnaires__post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Question": {
                "properties": {
                    "section_header": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Section Header"
                    },
                    "question_number": {
                        "type": "string",
                        "title": "Question Number"
                    },
                    "question_text": {
                        "type": "string",
                        "title": "Question Text"
                    },
                    "order_index": {
                        "type": "integer",
                        "title": "Order Index"
                    },
                    "assigned_to": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Assigned To"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "questionnaire_id": {
                        "type": "integer",
                        "title": "Questionnaire Id"
                    },
                    "project_id": {
                        "type": "integer",
                        "title": "Project Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "question_number",
                    "question_text",
                    "order_index",
                    "id",
                    "questionnaire_id",
                    "project_id",
                    "created_at"
                ],
                "title": "Question"
            },
            "Questionnaire": {
                "properties": {
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "project_id": {
                        "type": "integer",
                        "title": "Project Id"
                    },
                    "parsing_status": {
                        "$ref": "#/components/schemas/QuestionnaireStatus"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        },
                        "type": "array",
                        "title": "Questions",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "filename",
                    "id",
                    "project_id",
                    "parsing_status",
                    "version",
                    "created_at"
                ],
                "title": "Questionnaire"
            },
            "QuestionnaireStatus": {
                "type": "string",
                "enum": [
                    "PENDING",
                    "IN_PROGRESS",
                    "COMPLETED",
                    "FAILED"
                ],
                "title": "QuestionnaireStatus"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login/access-token"
                    }
                }
            }
        }
    }
}